pyproject.toml
[project]
name = "sauos"
version = "1.0.0"
description = "电脑全自动化系统 - Screen Automation Universal Operating System"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "SAUOS Team"}
]
keywords = ["automation", "desktop", "rpa", "gui", "screen"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Desktop Environment",
]

dependencies = [
    "pyautogui>=0.9.54",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pynput>=1.7.6",
    "mss>=9.0.0",
]

[project.optional-dependencies]
macos = [
    "pyobjc-framework-Quartz>=9.0",
]
linux = [
    "python-xlib>=0.33",
]
clipboard = [
    "pyperclip>=1.8.2",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
all = [
    "sauos[macos,linux,clipboard,dev]",
]

[project.scripts]
sauos = "sauos.__main__:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sauos*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]
